requirements.txt
Django>=3.1.5,<4.0
django-uuslug>=1.2.0,<2.0
Pillow>=8.0.1,<9.0
django-environ>=0.4.5,<1.0
django-manifest-loader>=1.0.0,<2.0
django-email-verification>=0.1.0,<1.0
django-redis>=4.12.1,<5.0
django-solo>=1.1.5,<2.0
django-admin-sortable>=2.2.3,<3.0
django-tinymce4-lite>=1.8.0,<2.0
six>=1.15.0,<2.0